Adds new allocate_k8s_token_v2 extension operation
Running on each control plane node is now an extension "server"
listening only on port 8800, rather than seprate Docker containers for
various extensions. The allocate_k8s_token_v2 extension will leverage
v2 of the extension server for the allocate_k8s_token operation. v2
returns not just a bootstrap token, but also the CA cert hash as well as
the API URL. In order to be able to migrate gradually to the v2
extension server for tokens, this new extension operation is being
created. Once the migration to v2 is complete we can look into retiring
the _v2 suffix and moving back to an extension operation with no suffix,
or possibly leaving the _v2 one and just removing the old "v1" url
mapping.